# performance optimization & code modernization