--- a/examples/chunk.rs
+++ b/examples/chunk.rs
@@ -15,7 +15,7 @@ fn my_default_predicate(x: u64) -> bool {
 }
 
 fn chunk_file<S: Into<String>>(path: S) -> io::Result<()> {
-    let f = try!(File::open(path.into()));
+    let f = File::open(path.into())?;
     let stream_length = f.metadata().unwrap().len();
     let reader: BufReader<File> = BufReader::new(f);
     let byte_iter = reader.bytes().map(|b| b.unwrap());