--- a/examples/tree01.rs
+++ b/examples/tree01.rs
@@ -17,15 +17,16 @@ fn my_new_node(level: usize, children: &Vec<IntHash>) -> Node<IntHash> {
     Node {
         hash: get_new_hash_id(),
         level: level,
-        children: children.clone()
+        children: children.clone(),
     }
 }
 
 fn main() {
     let levels = [0usize, 0, 1, 0, 1, 1, 2, 1, 0, 1, 0, 1];
     //let levels = [1usize, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1];
-    let hashed_chunk_it = levels.iter().enumerate().map(|(index, level)| {
-        HashedChunk {hash: index as IntHash, level: *level}
+    let hashed_chunk_it = levels.iter().enumerate().map(|(index, level)| HashedChunk {
+        hash: index as IntHash,
+        level: *level,
     });
 
     unsafe {