--- a/src/chunk.rs
+++ b/src/chunk.rs
@@ -1,4 +1,4 @@
-use super::{Separator};
+use super::Separator;
 
 pub struct Chunk {
     pub index: u64,
@@ -12,7 +12,7 @@ pub struct ChunkIter<Iter> {
     last_separator_index: u64,
 }
 
-impl<Iter: Iterator<Item=Separator>> ChunkIter<Iter> {
+impl<Iter: Iterator<Item = Separator>> ChunkIter<Iter> {
     pub fn new(iter: Iter, stream_length: u64) -> ChunkIter<Iter> {
         ChunkIter {
             separators: iter,
@@ -22,7 +22,7 @@ impl<Iter: Iterator<Item=Separator>> ChunkIter<Iter> {
     }
 }
 
-impl<Iter: Iterator<Item=Separator>> Iterator for ChunkIter<Iter> {
+impl<Iter: Iterator<Item = Separator>> Iterator for ChunkIter<Iter> {
     type Item = Chunk;
 
     fn next(&mut self) -> Option<Self::Item> {
@@ -35,7 +35,7 @@ impl<Iter: Iterator<Item=Separator>> Iterator for ChunkIter<Iter> {
                     size: chunk_size,
                     separator_hash: separator.hash,
                 });
-            },
+            }
             None => {
                 let chunk_size = self.stream_length - self.last_separator_index;
                 self.last_separator_index = self.stream_length;
@@ -45,8 +45,7 @@ impl<Iter: Iterator<Item=Separator>> Iterator for ChunkIter<Iter> {
                         size: chunk_size,
                         separator_hash: 0, // any value is ok, last chunk of the stream.
                     });
-                }
-                else {
+                } else {
                     return None;
                 }
             }