--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,11 +1,11 @@
+mod chunk;
 mod polynom;
 mod rolling_hash;
 mod separator;
-mod chunk;
 mod tree;
 
-pub use polynom::{Polynom, Polynom64};
-pub use rolling_hash::{RollingHash64, Rabin64};
-pub use separator::{Separator, SeparatorIter, HashToLevel};
 pub use chunk::{Chunk, ChunkIter};
+pub use polynom::{Polynom, Polynom64};
+pub use rolling_hash::{Rabin64, RollingHash64};
+pub use separator::{HashToLevel, Separator, SeparatorIter};
 pub use tree::{HashedChunk, Node, NodeIter};